plot: rename log into logf
Having a function named `log` is relatively annoying when one is
debugging, especially in conjunction with `goimports`:
adding `log.Fatalf("foo")` statements is confusing `goimports` as a
function with the name `log` is already defined.

Rename `log` into `logf`.

+++ axis.go
@@ -167,8 +167,8 @@ var _ Normalizer = LogScale{}
 // Normalize returns the fractional logarithmic distance of
 // x between min and max.
 func (LogScale) Normalize(min, max, x float64) float64 {
-	logMin := log(min)
-	return (log(x) - logMin) / (log(max) - logMin)
+	logMin := logf(min)
+	return (logf(x) - logMin) / (logf(max) - logMin)
 }
 
 // InvertedScale can be used as the value of an Axis.Scale function to
@@ -610,7 +610,7 @@ func tickLabelWidth(sty draw.TextStyle, ticks []Tick) vg.Length {
 	return maxWidth
 }
 
-func log(x float64) float64 {
+func logf(x float64) float64 {
 	if x <= 0 {
 		panic("Values must be greater than 0 for a log scale.")
 	}
